+++ packages/flutter_riverpod/test/listen_test.dart
@@ -8,7 +8,7 @@ import 'package:mockito/mockito.dart';
 import 'utils.dart';
 
 void main() {
-  group('WidgetRef.listenOnce', () {
+  group('WidgetRef.listenManual', () {
     testWidgets('listens to changes', (tester) async {
       final provider = StateProvider((ref) => 0);
       final listener = Listener<int>();
@@ -26,7 +26,7 @@ void main() {
         ),
       );
 
-      ref.listenOnce(provider, listener);
+      ref.listenManual(provider, listener);
 
       ref.read(provider.notifier).state++;
       verifyOnly(listener, listener(0, 1));
@@ -57,7 +57,7 @@ void main() {
         tester.element(find.byType(Consumer)),
       );
 
-      ref.listenOnce(provider, listener);
+      ref.listenManual(provider, listener);
 
       await tester.pumpWidget(ProviderScope(child: Container()));
 
@@ -81,7 +81,7 @@ void main() {
         ),
       );
 
-      ref.listenOnce(provider, listener, fireImmediately: true);
+      ref.listenManual(provider, listener, fireImmediately: true);
 
       verifyOnly(listener, listener(null, 0));
     });
@@ -102,7 +102,7 @@ void main() {
         ),
       );
 
-      final sub = ref.listenOnce(provider, (prev, next) {});
+      final sub = ref.listenManual(provider, (prev, next) {});
 
       expect(sub.read(), 0);
 
@@ -127,7 +127,7 @@ void main() {
         ),
       );
 
-      final sub = ref.listenOnce(provider, listener);
+      final sub = ref.listenManual(provider, listener);
 
       ref.read(provider.notifier).state++;
       verifyOnly(listener, listener(0, 1));